enums moved to proper location
