+++ api/src/main/java/cz/fi/muni/pa165/api/dto/CarDTO.java
@@ -5,7 +5,7 @@
  */
 package cz.fi.muni.pa165.api.dto;
 
-import cz.fi.muni.pa165.enums.CarState;
+import cz.fi.muni.pa165.api.enums.CarState;
 import java.util.Date;
 import java.util.Objects;
 import java.util.UUID;
