+++ api/src/main/java/cz/fi/muni/pa165/api/enums/CarState.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package cz.fi.muni.pa165.enums;
+package cz.fi.muni.pa165.api.enums;
 
 
 /**
