+++ api/src/main/java/cz/fi/muni/pa165/api/facade/ApplicationRejectedRecordFacadeImpl.java
@@ -32,6 +32,7 @@ public class ApplicationRejectedRecordFacadeImpl implements ApplicationRejectedR
 
     @Autowired
     private CarService carService;
+
     /*
     @Autowired
     private UserService userService;*/
