+++ api/src/main/java/cz/fi/muni/pa165/api/facade/CarFacadeImpl.java
@@ -7,7 +7,7 @@ package cz.fi.muni.pa165.api.facade;
 
 import cz.fi.muni.pa165.api.dto.CarDTO;
 import cz.fi.muni.pa165.model.entity.Car;
-import cz.fi.muni.pa165.enums.CarState;
+import cz.fi.muni.pa165.api.enums.CarState;
 import cz.fi.muni.pa165.service.BeanMappingService;
 import cz.fi.muni.pa165.service.CarService;
