+++ persistence/src/main/java/cz/fi/muni/pa165/model/DomainException.java
@@ -1,6 +1,6 @@
 package cz.fi.muni.pa165.model;
 
-import cz.fi.muni.pa165.enums.CarState;
+import cz.fi.muni.pa165.model.enums.CarState;
 
 /**
  * @author jkuchar
