+++ persistence/src/main/java/cz/fi/muni/pa165/model/dao/CarDao.java
@@ -1,6 +1,6 @@
 package cz.fi.muni.pa165.model.dao;
 
-import cz.fi.muni.pa165.enums.CarState;
+import cz.fi.muni.pa165.model.enums.CarState;
 import cz.fi.muni.pa165.model.entity.Car;
 
 import java.util.List;
