+++ persistence/src/main/java/cz/fi/muni/pa165/model/entity/Car.java
@@ -1,6 +1,6 @@
 package cz.fi.muni.pa165.model.entity;
 
-import cz.fi.muni.pa165.enums.CarState;
+import cz.fi.muni.pa165.model.enums.CarState;
 import cz.fi.muni.pa165.model.DomainException;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.util.Assert;
