+++ service/src/main/java/cz/fi/muni/pa165/service/CarService.java
@@ -6,7 +6,7 @@
 package cz.fi.muni.pa165.service;
 
 
-import cz.fi.muni.pa165.enums.CarState;
+import cz.fi.muni.pa165.model.enums.CarState;
 import cz.fi.muni.pa165.model.entity.Car;
 import java.util.List;
 import java.util.UUID;
